Flying ground ( Lift ) logic is now working
When mario jump on the lift, it will descend and bring mario together with it.
by using currentFlyingGround to tag the lift, it is easy for us to handle the pulling down activity by controlling it in liftUpdate() func
using Raycasting ( a mordern collision detector) to check that mario is on the fg or not.